Revert "cleaning up"
This reverts commit e19fa80f48f2f6542538c594290e74830caa967f.

+++ Robot_Arm/Servo_Joystick_PCA9685/Joint.cpp
@@ -1,6 +1,5 @@
 #include "Joint.hh"
 #include "HardwareSerial.h"
-
 #include <Adafruit_PWMServoDriver.h>
 #include <Arduino.h>
 #include <Wire.h>
@@ -10,8 +9,8 @@ extern HardwareSerial Serial;
 
 joint::joint() {}
 
-void joint::setJoint(uint8_t pinNum, uint16_t minVal, uint16_t maxVal,
-                     uint16_t pos, uint16_t startPos) {
+void joint::setJoint(byte pinNum, unsigned minVal, unsigned maxVal,
+                     unsigned pos, unsigned startPos) {
   this->pinNum = pinNum;
   this->maxVal = maxVal;
   this->minVal = minVal;
@@ -21,18 +20,32 @@ void joint::setJoint(uint8_t pinNum, uint16_t minVal, uint16_t maxVal,
 
 void joint::move() { pwm.writeMicroseconds(this->pinNum, this->pos); }
 
-void joint::moveMs(uint16_t time) {
+void joint::moveMs(unsigned time) {
   this->pos = time;
   pwm.writeMicroseconds(this->pinNum, this->pos);
 }
 
-uint8_t joint::moveDegree(uint8_t angle) {
+uint8_t joint::moveDegree(int angle) {
+  // Serial.print(angle);
+
+  // if (angle > 180)
+    // Serial.print("\t");
+    // Serial.println((pinNum + 7));
+  //   return ((pinNum + 1) * 10);
+  // if (angle < 0)
+    // Serial.print("\t");
+    // Serial.println(pinNum++);
+    // return pinNum++;
+
   this->angle = angle;
   if (maxVal > 1800) {
     this->pos = map(angle, 0, 180, minVal, maxVal);
   } else {
     this->pos = map(angle, 0, 90, minVal, maxVal);
   }
+
+  // Serial.print("\t");
+  // Serial.println(pos);
 
   move();
