+++ Robot_Arm/Servo_Joystick_PCA9685/Robot.hh
@@ -1,8 +1,6 @@
 #ifndef ROBOT_H
 #define ROBOT_H
-
 #include "Joint.hh"
-
 #include <Arduino.h>
 #include <stdint.h>
 
@@ -15,9 +13,9 @@ private:
   joint *wristRot;
   joint *gripper;
 
-  float xPos = 0;
-  float yPos = 350;
-  float zPos = 150;
+  int xPos = 0;
+  int yPos = 350;
+  int zPos = 150;
   int8_t grippingAngle = -90;
   int8_t rotionDegree = 90;
 
@@ -28,10 +26,10 @@ private:
   uint8_t y2Pin;
   uint8_t button2Pin;
 
-  uint16_t baseHight;
-  uint16_t shoulderToElbow;
-  uint16_t elbowToWrist;
-  uint16_t gripperLength;
+  unsigned baseHight;
+  unsigned shoulderToElbow;
+  unsigned elbowToWrist;
+  unsigned gripperLength;
 
 public:
   Robot();
@@ -43,22 +41,21 @@ public:
   void setDimension(unsigned baseHight, unsigned shoulderToElbow,
                     unsigned elbowToWrist, unsigned gripperLength);
 
-  void setJoystick(uint8_t x1Pin, uint8_t y1Pin, uint8_t button1Pin,
-                   uint8_t x2Pin, uint8_t y2Pin, uint8_t button2Pin);
+  void setJoystick(byte x1Pin, byte y1Pin, byte button1Pin, byte x2Pin,
+                   byte y2Pin, byte button2Pin);
 
   /* --- move --- */
   uint8_t moveEndEffector(float xVal, float yVal, float zVal,
                           int8_t grippingAngle);
 
-  void moveJoints(uint8_t baseAngle, uint8_t shoulderAngle, uint8_t elbowAngle,
-                  uint8_t wristAngle, uint8_t gripperAngle,
-                  uint8_t gripperOpening);
+  void moveJoints(byte baseAngle, byte shoulderAngle, byte elbowAngle,
+                  byte wristAngle, byte gripperAngle, byte gripperOpening);
 
   void moveWithJoystick();
 
   uint8_t moveEndEffector_Joystick();
 
-  /* --- Demos --- */
+  /* --- Demos ---*/
   uint8_t moveEndEffector_Demo();
 };
